Move logout to `auth` package
