+++ cli/internal/auth/logout.go
@@ -1,13 +1,13 @@
-package team
+package auth
 
 import (
 	"fmt"
-	"github.com/cloudquery/cloudquery-api-go/auth"
+	cqapiauth "github.com/cloudquery/cloudquery-api-go/auth"
 	"github.com/cloudquery/cloudquery-api-go/config"
 )
 
 func Logout() error {
-	err := auth.RemoveRefreshToken()
+	err := cqapiauth.RemoveRefreshToken()
 	if err != nil {
 		return fmt.Errorf("failed to remove refresh token: %w", err)
 	}
